Issue fix and presumed error fix
Fixed Issue: "The Portfolio option is not redirecting to the specified
link. Please fix. #4" - URL link to portfolio not working.

Also corrected unclosed divs in index.html - I presume this was a
mistake - please check, but if I am right (they should have been closed
within the section), not including this correction could lead to more
problems, and makes the HTML invalid.